(Update) Categories System
- Make UNIT3D more general friendly. On top of the icons from previous
commit there is now option to select if there will be meta for the
category or not. Movie and TV cats should have meta enabled. Music,
Apps and whatnot should not. If meta is not enabled for said category
then the meta block is hidden for said torrent or request on its
details page.