(Update) .env.example (read body)
You can now choose to store your api keys in the .env file as this
is a safer place for such sensitive data.

If you pull this commit into your project you will need to update
your api keys either in the api-keys.php or the .env file.

In addition we fixed a bug in the redis database configurations.

Any questions please visit us on discord!