(Alpha) Chat v2.0 (pusher only)
This is a "Alpha" commit. Its meant for Alpha testing. Features are
still being developed.

Please note, you must setup at least a FREE account with pusher and set
the corresponding fields in you `.env` file.

In the coming commits I will introducing a fallback so that once the
FREE pusher DAILY limits are reached, it will fall back to tradiational
database pulling every 3 seconds to get new messages.

There are still features that need to be added:
1. Emoji's
2. Statuses
3. User Tagging
4. Users List
5. Delete own messages, and staff moderation
and more ...