(feature) git:update artisan command
This feature adds a `php artisan git:update` command. Its used to
update the local repository with the current remote repository without
changing local files.

This is a Beta release!

Disclaimer: We are not responsible for any data lost with the usage of
this tool. Use at your own risk. Please report issues so that we can
improve on this tool.