(feature)[git:update] Beta v1.6
Can now update a single file.

Example: `php artisan git:update path/to/file/to/update`

Passing no argument does a full update, otherwise just a single file
update.

Disclaimer: We are not responsible for any data lost with the usage of
this tool. Use at your own risk. Please report issues so that we can
improve on this tool.