(Fix) Fixed status changed messages
Note: Requires to run `npm run prod`

This fixes an issue with the system user not being able to "post" on
behalf of the authorized user. Instead the authorized user will now
"post" his own status updates.

This also changes the look of the status update messages.

The status messages WILL NOT be presisted to the database, only
broadcasted to the current online chat users. This will allow for more
relevant chat history and help towards optimizing the overall
performance.