add: Replace "type" with "type-verb" @lang for certain pages.
All appear in hinting emoji typing.

It's hard for some languages' user to distinguish
verb or noun of word "type". This commit try to
deal with it. Feel free to cherry-pick / rebase
drop.

With this commit and the previous one, the lang
item "common.type" now is used as noun and
feature the same semantic as plural lang item
 "common.types".