feat: Reduce test suite execution time considerably
Given the sheer volume of migrations, it takes about 40 seconds on an average system to execute them all prior to running the test suite.

Loading one flat SQL file instead of executing each migration reduces this overhead significantly.

This is implemented in a way that degrades gracefully; if the flat file is not specified, all migrations will run as normal.